04_two_string_arrays_are_equivalence
LeetCode Problem of The Day (25 Oct 2022)  1662. Check If Two String Arrays are Equivalent                                                                                                                                                                                                          Given two string arrays word1 and word2, return true if the two arrays represent the same string, and false otherwise.
A string is represented by an array if the array elements concatenated in order forms the string.